% Auto-generated: do not edit by hand
\name{''DashMaplibre}

\alias{''DashMaplibre}

\title{DashMaplibre component}

\description{

}

\usage{
''DashMaplibre(id=NULL, basemap=NULL, bearing=NULL, center=NULL,
colorbar_map=NULL, colorbar_risk=NULL, hover_html=NULL,
hover_layer=NULL, layers=NULL, max_bounds=NULL, pitch=NULL,
sources=NULL, style=NULL, zoom=NULL)
}

\arguments{
\item{id}{Character. }

\item{basemap}{Character | named list. }

\item{bearing}{Numeric. }

\item{center}{Unnamed list. }

\item{colorbar_map}{Named list. }

\item{colorbar_risk}{Named list. }

\item{hover_html}{Character. }

\item{hover_layer}{Character. }

\item{layers}{Unnamed list. }

\item{max_bounds}{Unnamed list. }

\item{pitch}{Numeric. }

\item{sources}{Named list. }

\item{style}{Named list. }

\item{zoom}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

