% Auto-generated: do not edit by hand
\name{''DashMaplibre}

\alias{''DashMaplibre}

\title{DashMaplibre component}

\description{

}

\usage{
''DashMaplibre(id=NULL, bearing=NULL, center=NULL, colorbar_map=NULL,
colorbar_risk=NULL, layers=NULL, pitch=NULL, sources=NULL,
style=NULL, styleUrl=NULL, zoom=NULL)
}

\arguments{
\item{id}{Character. }

\item{bearing}{Numeric. }

\item{center}{Unnamed list. }

\item{colorbar_map}{Named list. }

\item{colorbar_risk}{Named list. }

\item{layers}{Unnamed list. }

\item{pitch}{Numeric. }

\item{sources}{Named list. }

\item{style}{Named list. }

\item{styleUrl}{Character. }

\item{zoom}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

